% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util__createInsertQuery.R
\name{createInsertQuery}
\alias{createInsertQuery}
\title{Create a SQL INSERT query string.}
\usage{
createInsertQuery(df, tblname, .append = "")
}
\arguments{
\item{df}{A data frame to insert into SQL.}

\item{tblname}{The name of the SQl table.}

\item{.append}{(string) Any additional characters to append to the end of the query.}
}
\value{
A character string representing a SQL query.
}
\description{
Create a SQL INSERT query string.
}
