% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util__mapTsCols.R
\name{mapTsCols}
\alias{mapTsCols}
\title{Iterates through the columns of a time-series tibble with a "date" column}
\usage{
mapTsCols(df, fn = function(x) x)
}
\description{
Iterates through the columns of a time-series tibble with a "date" column
}
