% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util___monthToQuarter.R
\name{monthToQuarter}
\alias{monthToQuarter}
\title{Aggregate a monthly data frame into a quarterly data frame}
\usage{
monthToQuarter(df, .requireAll = TRUE)
}
\arguments{
\item{df:}{A data frame with a date column and a value column}

\item{.requireAll:}{Boolean, if TRUE then will only return quarters for which there are 3 months of data}
}
\description{
Aggregate a monthly data frame into a quarterly data frame
}
