% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util__strdateToDate.R
\name{strdateToDate}
\alias{strdateToDate}
\title{Converts string dates of quarterly format (2012Q1, 2012Q2, ...) or monthly format to corresponding R Date objects.}
\usage{
strdateToDate(strdates)
}
\arguments{
\item{strdates}{A vector of string dates of quarterly or monthly format (YYYYQx or YYYYMxx).}
}
\value{
A vector a R date objects
}
\description{
Converts string dates of quarterly format (2012Q1, 2012Q2, ...) or monthly format to corresponding R Date objects.
}
